refactor: auto-split test_ball_roll_physics.py
