+++ tests/unit/engines/putting_green/ball_roll_physics/test_ball_roll_physics.py
@@ -19,101 +19,6 @@ from src.engines.physics_engines.putting_green.python.turf_properties import (
 )
 
 
-class TestRollMode:
-    """Tests for RollMode enumeration."""
-
-    def test_roll_modes_exist(self) -> None:
-        """Verify all roll modes are defined."""
-        assert hasattr(RollMode, "SLIDING")
-        assert hasattr(RollMode, "ROLLING")
-        assert hasattr(RollMode, "STOPPED")
-
-
-class TestBallState:
-    """Tests for BallState dataclass."""
-
-    def test_ball_state_creation(self) -> None:
-        """BallState should store position, velocity, spin."""
-        state = BallState(
-            position=np.array([5.0, 5.0]),
-            velocity=np.array([2.0, 0.0]),
-            spin=np.array([0.0, 0.0, 100.0]),  # Backspin about x-axis
-        )
-        assert np.allclose(state.position, [5.0, 5.0])
-        assert np.allclose(state.velocity, [2.0, 0.0])
-        assert state.spin[2] == 100.0
-
-    def test_ball_state_speed_property(self) -> None:
-        """Speed should be magnitude of velocity."""
-        state = BallState(
-            position=np.array([0.0, 0.0]),
-            velocity=np.array([3.0, 4.0]),
-            spin=np.zeros(3),
-        )
-        assert np.isclose(state.speed, 5.0)
-
-    def test_ball_state_column_vector_velocity_is_supported(self) -> None:
-        """Column-vector velocity input should be normalized to 1D."""
-        state = BallState(
-            position=np.array([[0.0], [0.0]]),
-            velocity=np.array([[3.0], [4.0]]),
-            spin=np.array([[0.0], [0.0], [0.0]]),
-        )
-        assert state.velocity.shape == (2,)
-        assert np.isclose(state.speed, 5.0)
-        assert np.allclose(state.direction, np.array([0.6, 0.8]))
-
-    def test_ball_state_row_vector_velocity_is_supported(self) -> None:
-        """Row-vector velocity input should be normalized to 1D."""
-        state = BallState(
-            position=np.array([[0.0, 0.0]]),
-            velocity=np.array([[3.0, 4.0]]),
-            spin=np.array([[0.0, 0.0, 0.0]]),
-        )
-        assert state.velocity.shape == (2,)
-        assert np.isclose(state.speed, 5.0)
-
-    def test_ball_state_is_moving(self) -> None:
-        """Should detect if ball is moving."""
-        moving = BallState(
-            position=np.array([0.0, 0.0]),
-            velocity=np.array([1.0, 0.0]),
-            spin=np.zeros(3),
-        )
-        stopped = BallState(
-            position=np.array([0.0, 0.0]),
-            velocity=np.array([0.0, 0.0]),
-            spin=np.zeros(3),
-        )
-        assert moving.is_moving
-        assert not stopped.is_moving
-
-    def test_ball_state_copy(self) -> None:
-        """Should create independent copy."""
-        original = BallState(
-            position=np.array([1.0, 1.0]),
-            velocity=np.array([2.0, 2.0]),
-            spin=np.array([0.0, 0.0, 50.0]),
-        )
-        copy = original.copy()
-
-        # Modify copy
-        copy.position[0] = 999.0
-
-        # Original should be unchanged
-        assert original.position[0] == 1.0
-
-    def test_ball_state_direction(self) -> None:
-        """Should compute unit direction vector."""
-        state = BallState(
-            position=np.array([0.0, 0.0]),
-            velocity=np.array([3.0, 4.0]),
-            spin=np.zeros(3),
-        )
-        direction = state.direction
-        assert np.isclose(np.linalg.norm(direction), 1.0)
-
-
 class TestBallRollPhysics:
     """Tests for BallRollPhysics class."""
 
@@ -403,139 +308,3 @@ class TestBallRollPhysics:
 
         # Ball with backspin should have traveled less
         assert with_backspin.position[0] < without_backspin.position[0]
-
-
-class TestBallRollPhysicsEdgeCases:
-    """Edge case tests for BallRollPhysics."""
-
-    @pytest.fixture
-    def physics(self) -> BallRollPhysics:
-        return BallRollPhysics()
-
-    def test_zero_velocity_step(self, physics: BallRollPhysics) -> None:
-        """Step with zero velocity should not move ball."""
-        state = BallState(
-            position=np.array([5.0, 5.0]),
-            velocity=np.array([0.0, 0.0]),
-            spin=np.zeros(3),
-        )
-
-        new_state = physics.step(state, dt=0.1)
-
-        # Position should be unchanged
-        assert np.allclose(new_state.position, state.position)
-
-    def test_very_small_velocity_stops(self, physics: BallRollPhysics) -> None:
-        """Very small velocity should snap to zero (stopping threshold)."""
-        state = BallState(
-            position=np.array([5.0, 5.0]),
-            velocity=np.array([0.001, 0.0]),  # Very slow
-            spin=np.zeros(3),
-        )
-
-        new_state = physics.step(state, dt=0.1)
-
-        # Should stop
-        assert not new_state.is_moving
-
-    def test_large_dt_stability(self, physics: BallRollPhysics) -> None:
-        """Large time steps should not cause numerical instability."""
-        state = BallState(
-            position=np.array([5.0, 5.0]),
-            velocity=np.array([2.0, 0.0]),
-            spin=np.zeros(3),
-        )
-
-        # Large time step
-        new_state = physics.step(state, dt=1.0)
-
-        # Should be finite
-        assert np.all(np.isfinite(new_state.position))
-        assert np.all(np.isfinite(new_state.velocity))
-
-    def test_negative_velocity(self, physics: BallRollPhysics) -> None:
-        """Negative velocity should work correctly."""
-        state = BallState(
-            position=np.array([10.0, 10.0]),
-            velocity=np.array([-2.0, -1.0]),
-            spin=np.zeros(3),
-        )
-
-        new_state = physics.step(state, dt=0.1)
-
-        # Should move in negative direction
-        assert new_state.position[0] < state.position[0]
-        assert new_state.position[1] < state.position[1]
-
-    def test_extreme_spin(self, physics: BallRollPhysics) -> None:
-        """Extreme spin values should be handled."""
-        state = BallState(
-            position=np.array([5.0, 5.0]),
-            velocity=np.array([2.0, 0.0]),
-            spin=np.array([0.0, 10000.0, 0.0]),  # Extreme spin
-        )
-
-        # Should not crash
-        new_state = physics.step(state, dt=0.01)
-        assert np.all(np.isfinite(new_state.velocity))
-
-    def test_timestep_independence(self, physics: BallRollPhysics) -> None:
-        """Results should be similar regardless of timestep size (within reason)."""
-        initial = BallState(
-            position=np.array([5.0, 5.0]),
-            velocity=np.array([2.0, 0.0]),
-            spin=np.zeros(3),
-        )
-
-        # Simulate with small timesteps
-        state_small = initial.copy()
-        for _ in range(100):
-            state_small = physics.step(state_small, dt=0.001)
-
-        # Simulate with larger timesteps
-        state_large = initial.copy()
-        for _ in range(10):
-            state_large = physics.step(state_large, dt=0.01)
-
-        # Should be reasonably close (not exact due to numerical differences)
-        assert np.allclose(state_small.position, state_large.position, rtol=0.1)
-
-
-class TestBallRollPhysicsIntegrators:
-    """Tests for different integration methods."""
-
-    def test_euler_integration(self) -> None:
-        """Euler integration should work."""
-        physics = BallRollPhysics(integrator="euler")
-        state = BallState(
-            position=np.array([0.0, 0.0]),
-            velocity=np.array([2.0, 0.0]),
-            spin=np.zeros(3),
-        )
-
-        new_state = physics.step(state, dt=0.01)
-        assert np.all(np.isfinite(new_state.position))
-
-    def test_rk4_integration(self) -> None:
-        """RK4 integration should be more accurate."""
-        physics = BallRollPhysics(integrator="rk4")
-        state = BallState(
-            position=np.array([0.0, 0.0]),
-            velocity=np.array([2.0, 0.0]),
-            spin=np.zeros(3),
-        )
-
-        new_state = physics.step(state, dt=0.01)
-        assert np.all(np.isfinite(new_state.position))
-
-    def test_verlet_integration(self) -> None:
-        """Verlet integration for better energy conservation."""
-        physics = BallRollPhysics(integrator="verlet")
-        state = BallState(
-            position=np.array([0.0, 0.0]),
-            velocity=np.array([2.0, 0.0]),
-            spin=np.zeros(3),
-        )
-
-        new_state = physics.step(state, dt=0.01)
-        assert np.all(np.isfinite(new_state.position))
